third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/script-charset-01-expected.txt

This is a testharness.js-based test.
[FAIL] Script @type: unknown parameters 3
  assert_equals: expected "śćążź" but got "\\ufffd湿\\ufffd"
Harness: the test ran to completion.
